Docker container setup and runs with cassandra
