+++ .docker/database/apt/general.apt
@@ -1,0 +1,5 @@
+apt-utils
+iputils-ping
+tmux
+vim
+net-tools
