+++ .docker/database/run/cqlsh
@@ -1,0 +1,39 @@
+; Licensed to the Apache Software Foundation (ASF) under one
+; or more contributor license agreements.  See the NOTICE file
+; distributed with this work for additional information
+; regarding copyright ownership.  The ASF licenses this file
+; to you under the Apache License, Version 2.0 (the
+; "License"); you may not use this file except in compliance
+; with the License.  You may obtain a copy of the License at
+;
+;   http://www.apache.org/licenses/LICENSE-2.0
+;
+; Unless required by applicable law or agreed to in writing,
+; software distributed under the License is distributed on an
+; "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+; KIND, either express or implied.  See the License for the
+; specific language governing permissions and limitations
+; under the License.
+;
+; Sample ~/.cqlshrc file.
+
+[authentication]
+;; If Cassandra has auth enabled, fill out these options
+; username = fred
+; password = !!bang!!$
+; keyspace = ks1
+
+[ui]
+;; Whether or not to display query results with colors
+color = on
+
+;; Used for displaying timestamps (and reading them with COPY)
+; datetimeformat = %Y-%m-%d %H:%M:%S%z
+
+;; Display timezone
+;timezone = Etc/UTC
+
+;; The number of digits displayed after the decimal point for single and double precision numbers
+;; (note that increasing this to large numbers can result in unusual values)
+float_precision = 30
+double_precision = 30
