+++ .env
@@ -1,0 +1,10 @@
+# Project vars
+PROJECT_NAME=apt_decoder
+IMAGE_PREFIX=apt_decoder
+SUBNET_BASE=172.23.0
+## Add your dockerhub username below
+REGISTRY_URL=superelectron
+
+# Database vars
+DB_USER=cassandra
+DB_PWD=cassandra
